dealership/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-environ==0.4.5
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
isort==5.8.0
Markdown==3.3.4
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
